requirements.txt
# BioArt Learning Process Dependencies

# Core data processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Bioinformatics and sequence analysis
biopython>=1.79

# Machine learning and embeddings
scikit-learn>=1.0.0
torch>=1.11.0
transformers>=4.18.0

# Computer vision and image processing
opencv-python>=4.5.0
pillow>=8.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Generative models (optional, for advanced users)
diffusers>=0.10.0  # For diffusion models
peft>=0.3.0  # For LoRA fine-tuning

# Data downloading and management
kaggle>=1.5.12
requests>=2.25.0
tqdm>=4.62.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Note: This project focuses on computational bioart, not biological function
# All dependencies are for creative/educational purposes only